turn applications context into behaviour
